refined the checkout controller flows and methods
