+++ spring-backend/src/main/resources/templates/checkout.html
@@ -1,35 +1,35 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Stripe Checkout</title>
-    <script src="https://js.stripe.com/v3/"></script>
-</head>
-<body>
-<button id="checkout-button">Pay Now</button>
-
-<script>
-    const stripe = Stripe("pk_test_51RBtJ5GBEBEiOSMS84fM0Eozu7inX60wlw19XZ8RmsFlXDdyYIgINu1R8BsdCKfD2o3v6TEGUEDqtC0N0KPZdVAd00BGENAvN7"); // Replace with your Stripe publishable key
-
-    document.getElementById("checkout-button").addEventListener("click", () => {
-        fetch("http://localhost:8080/api/payment/create-checkout-session", { // Replace with your backend URL
-            method: "POST",
-            headers: { "Content-Type": "application/json" },
-        })
-            .then(response => response.json())
-            .then(data => {
-                return stripe.redirectToCheckout({ sessionId: data.sessionId });
-            })
-            .then(result => {
-                if (result.error) {
-                    alert(result.error.message);
-                }
-            })
-            .catch(error => {
-                console.error("Error:", error);
-            });
-    });
-</script>
-</body>
+	<head>
+	    <meta charset="UTF-8">
+	    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+	    <title>Stripe Checkout</title>
+	    <script src="https://js.stripe.com/v3/"></script>
+	</head>
+	<body>
+		<button id="checkout-button">Pay Now</button>
+		
+		<script>
+		    const stripe = Stripe("pk_test_51RBtJ5GBEBEiOSMS84fM0Eozu7inX60wlw19XZ8RmsFlXDdyYIgINu1R8BsdCKfD2o3v6TEGUEDqtC0N0KPZdVAd00BGENAvN7"); // Replace with your Stripe publishable key
+		
+		    document.getElementById("checkout-button").addEventListener("click", () => {
+		        fetch("http://localhost:8080/api/cart/submit-order", { // Replace with your backend URL
+		            method: "POST",
+		            headers: { "Content-Type": "application/json" },
+		        })
+		            .then(response => response.json())
+		            .then(data => {
+		                return stripe.redirectToCheckout({ sessionId: data.sessionId });
+		            })
+		            .then(result => {
+		                if (result.error) {
+		                    alert(result.error.message);
+		                }
+		            })
+		            .catch(error => {
+		                console.error("Error:", error);
+		            });
+		    });
+		</script>
+	</body>
 </html>
